Comply AssetDatabaseUtility with StyleCop
